SRV. # Model.Order/Item (Order + userId, Items ~ price.FLOAT)

+++ server/src/models/OrderItemModel.ts
@@ -36,7 +36,7 @@ class OrderItemModel extends Model<
           allowNull: false,
         },
         name: { type: DataTypes.STRING, allowNull: false },
-        price: { type: DataTypes.DECIMAL(10, 2), allowNull: false },
+        price: { type: DataTypes.FLOAT, allowNull: false },
         quantity: { type: DataTypes.INTEGER, allowNull: false },
         orderId: { type: DataTypes.INTEGER, allowNull: false },
       },
